add listing commands
Added removeserver to utilize the list of unknown servers with keywords
when the keywords command is used in DMs. Permitted unfollow to be used
in DMs with channel IDs for the same reason.

Added a success helper similar to the error and question helpers.

Ran rustfmt.

Fixed a typo in the query used for follows.

Implemented std::error::Error for error type.

Logged slightly more information when reporting errors.