list keywords in dms
Added the ability to list keywords in DMs (which was planned when the
command was added, but not implemented until now).

Switched to a "bleeding edge" version of serenity that supports setting
allowed mentions to scrub all mentions from messages containing user
input when sent to a guild channel.

Ran rustfmt.