change keyword system
This changes keywords to be added to the entire server by default,
instead of requiring users to follow channels they want to keep up with.
User configuration options are still provided through channel-specific
keywords and the ability to mute channels to prevent notifications about
server-wide keywords there.

Because of the added complexity posed by this system, the commands
module was refactored into several submodules.

Started explicitly passing the gateway intents that highlights requires
to perhaps have a small performance benefit.

Renamed the removeserver command to remove-server because squishedcase
is ugly.