delete notifications of deleted messages
Added a system to record notifications so they can be deleted when the
original message is deleted. Also delete notifications when the original
message is edited to not include the keyword anymore, and edit the
notification to reflect the current message contents if an edited
message still contains the keyword.